Add `on_reconnect` callback on the async client
Async clients that are using the reconnect feature can now provide an
optional callback that will be invoked before the server is reconnected.
You may optionally update the address or auth settings in this callback
and they will be applied when the client reconnects.

This is useful for updating auth tokens which may have expired by the
time a reconnection is required.