Serialize the reconnect tests
Tests that rely on the reconnect socket server cannot run in parallel
because they ask the server to disconnect and restart as part of the
test. This can cause other tests running in parallel to fail in
mysterious ways.

By using the `serial_test` module and applying it selectively to the
affected tests, we can ensure that they are never executed concurrently.