Async reconnections (#400)
* Add reconnect settings to async builder

The builder can now be configured with the following properties to
control automatic reconnections on network errors or server disconnects:

- reconnect
- reconnect_on_disconnect
- reconnect_delay (min and max)
- max_reconnect_attempts

* Introduce `DisconnectReason` enum

The enum replaces the need for multiple `AtomicBool`'s to maintain the
disconnection reason. This makes the code easier to read and more
ergonomic to maintain the state.

* Serialize the reconnect tests

Tests that rely on the reconnect socket server cannot run in parallel
because they ask the server to disconnect and restart as part of the
test. This can cause other tests running in parallel to fail in
mysterious ways.

By using the `serial_test` module and applying it selectively to the
affected tests, we can ensure that they are never executed concurrently.