Fixed all clippy warnings.
There were many warnings causing the `make clippy` command to fail.
Some of the more noticable changes are moving the client.rs code
to the corresponding mod.rs and deleting the former. This was to
stop the clippy module inception warning. All unit tests, doc tests,
and examples are passing.