ci: fix most of the tests
There's still an error with ci.
engineio/src/lib.rs/test/tls_connector doesn't function the same as native-tls one (as it properly validates hostnames)

Needs further investigation as to why the hostname shows as invalid in ci. If we no longer need that, then this PR can be merged.