feat(socketio): send close reason when emitting Event::Close
Previously, when we emitted an Event::Close, we didn't specify a
reason for why this event was emitted. When implementing the close
notification on unexpected transport closes, we started sending
'transport close' as the payload for the close event.

We now decided to do this more consistently and added a close reason
when emitting the Event::Close, so that the client better knows
why this event was emitted.